HW14 Moved the service configuration into a separate package
